Move code to `/src` folder
