+++ webpack.config.client.js
@@ -2,7 +2,7 @@ const path = require('path');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 
 const config = {
-  entry: ['client.js'],
+  entry: ['./src/client.js'],
   output: {
     path: path.join(__dirname, 'build/'),
     filename: 'client.js',
